Refactored code with packages
