+++ src/main/java/duke/models/Event.java
@@ -1,4 +1,4 @@
-package models;
+package duke.models;
 
 public class Event extends Task {
