+++ src/main/java/duke/models/Task.java
@@ -1,4 +1,4 @@
-package models;
+package duke.models;
 
 public class Task {
     protected String description;
