+++ src/main/java/duke/utils/Commands.java
@@ -1,4 +1,4 @@
-package utils;
+package duke.utils;
 
 public enum Commands {
     BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN
