+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ Hello from
 |____/ \__,_|_|\_\___|
 
 ---->---->---->---->---->---->---->---->---->---->
-Hello! Imma Duke!
+Hello! Imma duke.Duke!
  What can I do for you?
 <----<----<----<----<----<----<----<----<----<----
 ---->---->---->---->---->---->---->---->---->---->
@@ -25,5 +25,5 @@ Here are the tasks in your list:
 2. [E] [ ] project meeting (at: Mon 2-4pm)
 <----<----<----<----<----<----<----<----<----<----
 ---->---->---->---->---->---->---->---->---->---->
-Bye. Duke misses you.
+Bye. duke.Duke misses you.
 <----<----<----<----<----<----<----<----<----<----
